requirements.txt
faster-whisper>=1.0.0
av>=11.0.0
huggingface-hub>=0.20.0
tqdm>=4.66.0
google-genai>=1.0.0
pydantic>=2.0.0
